requirements.txt
spacy==3.7.2
confection==0.1.4
vllm==0.4.1
transformers==4.46.0
fschat==0.2.36
openai==1.25.1
torchvision==0.17.1
sentence-transformers==2.7.0
matplotlib==3.8.4
accelerate==0.30.0
datasketch
pandas
boto3
bpe